pyproject.toml
[project]
name = "smartresumeassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "pypdf2==3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart==0.0.6",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
